core: wifi: Segregate WifiManager to it's own file
